+++ Codes/subgroup3/UC_09_ShopkeeperOrder/OrderListMaker.py
@@ -1,22 +1,6 @@
 import datetime as dt
 import OrderSender as ods
-
-class OrderList:
-    def __init__(self,item_name,item_brand, item_price, item_number):
-        self.item_name = item_name
-        self.item_brand = item_brand
-        self.item_price = item_price
-        self.item_number = item_number
-        self.total_price = 0
-        self.length = len(self.item_name)
-        for i in range(len(item_price)):
-            self.total_price += item_price[i]*item_number[i]
-        
-    def printOrderList(self):
-        print("OrderList>>.>>>>>")
-        for i in range(self.length):
-            print("{0: <20}{1: <20}{2: <20}{3: <20}".format(self.item_name[i],self.item_brand[i],self.item_price[i],self.item_number[i]))
-
+import OrderList as ol
 
 class OrderListMaker:
     def __init__(self, connected_wh):
@@ -79,7 +63,7 @@ class OrderListMaker:
             else:
                 print("invalid input : Out of stock!")
 
-        self.OrderList = OrderList(temp_name,temp_brand, temp_price, temp_number)
+        self.OrderList = ol.OrderList(temp_name,temp_brand, temp_price, temp_number, self.OrderNumberMaker())
         self.total_price = self.OrderList.total_price
 
     def check_stock(self, name, brand, number):
